requirements.txt
discord.py[voice]>=2.3.0
yt-dlp>=2023.10.13
python-dotenv>=1.0.0
PyNaCl>=1.5.0
aiohttp>=3.9.0
pytube
youtube-search-python